# docs/swagger.yaml
basePath: /
definitions:
  web.LoginRequest:
    properties:
      password:
        example: admin
        type: string
      username:
        example: admin
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for controlling audio interfaces, recording, and Gemini models.
  title: Abel API
  version: "1.0"
paths:
  /api/ai/config:
    get:
      description: Returns the list of configured languages and the original language
      produces:
      - application/json
      responses:
        "200":
          description: AI Configuration
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Get AI configuration
      tags:
      - AI
  /api/ai/streams:
    get:
      description: Returns active sessions and master state
      produces:
      - application/json
      responses:
        "200":
          description: AI Streams Status
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Get AI streams status
      tags:
      - AI
    post:
      consumes:
      - application/json
      description: Toggles master AI switch or stops a specific language translation
      parameters:
      - description: AI Action
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Invalid Action
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      security:
      - CookieAuth: []
      - BasicAuth: []
      summary: Control AI streams
      tags:
      - AI
  /api/ai/subtitles:
    get:
      description: Real-time SSE stream of subtitles for a specific language
      parameters:
      - description: Language
        in: query
        name: lang
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE Stream
          schema:
            type: string
      summary: Get subtitles stream
      tags:
      - AI
  /api/audio/config:
    get:
      description: Returns current channels, boost, device info, and recording status
      produces:
      - application/json
      responses:
        "200":
          description: Audio Configuration
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Get audio config
      tags:
      - Audio
    patch:
      consumes:
      - application/json
      description: Boots up the device engine or updates running config
      parameters:
      - description: Interface Config
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Invalid Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Error
          schema:
            type: string
      security:
      - CookieAuth: []
      - BasicAuth: []
      summary: Update audio config
      tags:
      - Audio
  /api/auth/session:
    post:
      consumes:
      - application/json
      description: Logs in and establishes a session
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/web.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login Success
          schema:
            type: object
      summary: Create auth session
      tags:
      - Auth
  /api/recordings:
    get:
      description: Returns whether the system is recording
      produces:
      - application/json
      responses:
        "200":
          description: Recording Status
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Get recording status
      tags:
      - Recordings
    post:
      consumes:
      - application/json
      description: Starts or stops a recording
      parameters:
      - description: Recording Action
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Invalid Action
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Error
          schema:
            type: string
      security:
      - CookieAuth: []
      - BasicAuth: []
      summary: Control recording
      tags:
      - Recordings
  /api/system/changelog:
    get:
      description: Real-time SSE stream of state changes across the console
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE Stream
          schema:
            type: string
      security:
      - CookieAuth: []
      - BasicAuth: []
      summary: Get system change log
      tags:
      - System
securityDefinitions:
  BasicAuth:
    type: basic
  CookieAuth:
    in: cookie
    name: abel_session
    type: apiKey
swagger: "2.0"